Transpiler: cross-file member typing + receiver-class resolution
Types class members across files and resolves member types through
receivers the previous inference gave up on — the machinery behind
draining the ORM decimal-into-int-member warning pile.

Member-type registry (hbreftab.c Pass 0, split into 0a mark-all-classes
then 0b register-members): each CLASS VAR's AS type becomes a
`Class::member` reftab row (namespace distinct from `Class__Method`
keys). SEND inference (hbtypes.c) and the emitter (gencsharp.c
hb_csSendMemberIsInteger) consult it along the INHERIT chain, so a
cross-file `oTransaction:nSaleClkNo` reads its declared AS INTEGER
instead of the Hungarian prefix. Param slots are seeded into the
emitter's local-type map so receivers arriving as parameters
(SetTableHeader(oTransaction, ...)) type too; the (long) write coercion
covers AS INTEGER members and ORM int/long fields alike.

Sink inference (Pass 2.5, HB_INT_MARK_SINK): a LOCAL whose value flows
into an INTEGER sink (an AS INTEGER member or ORM int field write) is
auto-typed INTEGER — zero source change. Gated by the same disqualifiers
as index candidacy, so a genuinely fractional value stays decimal.
hb_astExprIsIntegral gained an HB_ET_SEND case so int field/member READS
count as integral (they used to false-soft-disqualify the feeding local).

Class-typed members + nested receiver resolution: `VAR oClerk AS OBJECT`
name-seeds to class Clerk (the o+ClassName convention W0025 enforces),
GUARDED by hb_refTabIsClass(stem) — without the guard oPLU->"PLU" and
oOrmTable->"OrmTable" propagate as bogus receiver types and blow up the
build. A lean depth-capped hb_astResolveRecvClass then resolves nested
chains (oPOSStatus:oClerk:nNo) with reftab hash lookups only — NOT the
full hb_astInferExprType, whose recursive form once took scan from 0.08s
to 60s. Canary files stay at 0.03-0.07s.

HB_DEFAULT param typing: `hb_default(@param, ConstructORMTable(XxxDef()))`
— the idiom ORM helpers use to construct their table param in-body — now
types the param (routed through hb_astPropagateVar, so it only upgrades
an OBJECT/USUAL param, never overrides a caller class). This is the big
lever: it clears the field reads in ~18 helper bodies that otherwise
left the param generic OBJECT.

W0029 demoted to audit-only: a plain decimal-into-int-member/field write
COMPILES (the emitter inserts the (long) cast), so per the channel rule
(warning = definite fix that reaches zero; these don't — intended
truncations persist) it belongs in the ORM-NARROW audit, not the scan
gate. The two plain-assign fprintfs are removed, their audit rows kept
(the truncation signal survives for deliberate hunting); the compound
/= %= ^= case stays a warning (that one is a real CS0266). Corpus scan
warnings collapse to single digits.

verify.sh green 87/87 + 6 throughout.

+++ src/transpiler/tests/csout/test85.cs
@@ -70,7 +70,7 @@ public static partial class Program
         // through-base copy: branch takes 5/Alpha
         Stamp(oBranch, oDept);
         // then diverges: 9
-        oBranch.nNo = oBranch.nNo + 4;
+        oBranch.nNo = (long)(oBranch.nNo + 4);
         oBranch.cName = "Beta";
         oDept.nClerkNo = 9999999999;
         oBranch.nRegion = 42;
